chore(frontend): drop unused deps, dead code, and duplicate monitor SSE connection
Remove the unused @xyflow/react and elkjs dependencies from packages/monitor
and refresh the lockfile.

Lift the session detail event stream into a single EventSource in
SessionDetail and pass the parsed events, streaming state, and a clear
handler down to EventStream as props. Previously SessionDetail opened one
connection and each rendered EventStream opened another to the same /events
endpoint, so a session detail view held several connections at once; it now
holds exactly one while preserving event parsing, reconnect handling, and
what each consumer renders.

Remove wizard dead code: a leftover config setter console.log, the unused
is-returning route and its mock handler, and commented-out type aliases in
JsonUtils.
