+++ packages/monitor/package.json
@@ -16,8 +16,6 @@
   "dependencies": {
     "@fontsource/poppins": "^5.2.7",
     "@metalbear/ui": "^0.1.5",
-    "@xyflow/react": "^12.10.2",
-    "elkjs": "^0.11.1",
     "lucide-react": "^0.577.0",
     "posthog-js": "^1.364.6",
     "react": "^18.3.1",
