+++ packages/monitor/src/components/EventStream.tsx
@@ -1,29 +1,22 @@
 import { useState, useEffect, useRef } from "react";
 import { Button, Separator } from "@metalbear/ui";
 import { Activity, Trash2 } from "lucide-react";
-import type { SessionInfo, MonitorEvent } from "../types";
+import type { MonitorEvent, TimestampedEvent } from "../types";
 import type { EventTypeValue } from "../eventTypes";
 import { strings } from "../strings";
-import { api } from "../api";
 import EventFilterChips from "./events/EventFilterChips";
 import EventSearchBar from "./events/EventSearchBar";
 import EventRow from "./events/EventRow";
 import EventDetailDialog from "./events/EventDetailDialog";
-import { MAX_EVENTS } from "./events/eventConfig";
 import { parseEvent, type ParsedEvent } from "./events/parseEvent";
 
-interface TimestampedEvent {
-  event: MonitorEvent;
-  receivedAt: Date;
-}
-
 interface Props {
-  session: SessionInfo;
+  events: TimestampedEvent[];
+  streaming: boolean;
+  onClear: () => void;
 }
 
-export default function EventStream({ session }: Props) {
-  const [events, setEvents] = useState<TimestampedEvent[]>([]);
-  const [streaming, setStreaming] = useState(false);
+export default function EventStream({ events, streaming, onClear }: Props) {
   const [detailEvent, setDetailEvent] = useState<{
     summary: string;
     raw: string;
@@ -31,38 +24,6 @@ export default function EventStream({ session }: Props) {
   const [searchQuery, setSearchQuery] = useState("");
   const [activeFilter, setActiveFilter] = useState<EventTypeValue | null>(null);
   const logRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    setEvents([]);
-    setStreaming(true);
-
-    const eventSource = new EventSource(api.eventStreamUrl(session.session_id));
-
-    eventSource.onmessage = (e) => {
-      let event: MonitorEvent;
-      try {
-        event = JSON.parse(e.data);
-      } catch {
-        return;
-      }
-      setEvents((prev) => {
-        // Append the new event and cap the buffer at MAX_EVENTS by dropping
-        // the oldest entries. Keeps memory bounded for long-running sessions.
-        const next = [...prev, { event, receivedAt: new Date() }];
-        return next.length > MAX_EVENTS ? next.slice(-MAX_EVENTS) : next;
-      });
-    };
-
-    eventSource.onerror = () => {
-      setStreaming(false);
-      eventSource.close();
-    };
-
-    return () => {
-      eventSource.close();
-      setStreaming(false);
-    };
-  }, [session.session_id]);
 
   const isNearBottom = useRef(true);
   useEffect(() => {
@@ -146,7 +107,7 @@ export default function EventStream({ session }: Props) {
           <Button
             variant="ghost"
             size="icon"
-            onClick={() => setEvents([])}
+            onClick={onClear}
             title={strings.events.clear}
             className="h-6 w-6"
           >
